Revert "Merge remote-tracking branch 'origin/henrique' into abner"
This reverts commit 88367a58ae0112546129ddaaa2eb207055a8eb3f, reversing
changes made to 8efa67b15712c37864eaa4ac9ac938f4c45c878e.

+++ Assets/Scripts/Upgrade.cs
@@ -5,7 +5,7 @@ using System.Collections.Generic;
 public class Upgrade
 {
     public int[] cost;
-    public float[] value;
+    public int[] value;
 }
 
 public class Upgrades
